Split out compaction-task-creation module
