+++ java/compaction/compaction-task-creation/src/main/java/sleeper/compaction/taskcreation/RunTasks.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.jobexecution;
+package sleeper.compaction.taskcreation;
 
 import com.amazonaws.services.autoscaling.AmazonAutoScaling;
 import com.amazonaws.services.ecs.AmazonECS;
